backend/requirements.txt
# Django 프레임워크
Django>=4.2,<5.0

# Django REST Framework (API용)
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# MySQL 데이터베이스 연결
mysqlclient>=2.2.0

# 프로덕션 서버
gunicorn>=21.2.0

# 보안 관련
django-ratelimit>=4.1.0
django-axes>=6.1.0

# 인증 관련
djangorestframework-simplejwt>=5.3.0

# 파일 업로드
Pillow>=10.1.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 유틸리티
django-filter>=23.3

# 중첩 라우터
drf-nested-routers>=0.93.4
